Quality of Life Changes
Added a scrollTop to save where the user last left off at. Felt like before, it was too much of a hassle of opening new tabs for each piece and not being able to go back using the back button for browsers. Now users can open pages in new tabs if they would like to. Or just use the back button now.